# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JoinRequestGender =
  (typeof JoinRequestGender)[keyof typeof JoinRequestGender];

export const JoinRequestGender = {
  Male: "Male",
  Female: "Female",
} as const;

export interface JoinRequest {
  /**
   * @minLength 1
   * @maxLength 20
   */
  username: string;
  gender: JoinRequestGender;
}

export interface JoinResponse {
  userId: string;
  username: string;
  gender: string;
}

export interface OnlineUser {
  userId: string;
  username: string;
  gender: string;
  joinedAt: string;
}

export interface OnlineUsersResponse {
  users: OnlineUser[];
  total: number;
}

export interface Message {
  id: number;
  sessionId: string;
  fromUserId: string;
  toUserId: string;
  fromUsername: string;
  content: string;
  createdAt: string;
}

export interface MessagesResponse {
  messages: Message[];
}

export interface SendMessageRequest {
  sessionId: string;
  fromUserId: string;
  toUserId: string;
  fromUsername: string;
  /** @maxLength 1000 */
  content: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetOnlineUsersParams = {
  /**
   * Filter by gender
   */
  gender?: GetOnlineUsersGender;
};

export type GetOnlineUsersGender =
  (typeof GetOnlineUsersGender)[keyof typeof GetOnlineUsersGender];

export const GetOnlineUsersGender = {
  Male: "Male",
  Female: "Female",
  all: "all",
} as const;