# lib/api-zod/src/generated/types/joinRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { JoinRequestGender } from "./joinRequestGender";

export interface JoinRequest {
  /**
   * @minLength 1
   * @maxLength 20
   */
  username: string;
  gender: JoinRequestGender;
}