# lib/api-zod/src/generated/types/joinResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinResponse {
  userId: string;
  username: string;
  gender: string;
}