# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  sessionId: string;
  fromUserId: string;
  toUserId: string;
  fromUsername: string;
  content: string;
  createdAt: Date;
}