# lib/api-zod/src/generated/types/onlineUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface OnlineUser {
  userId: string;
  username: string;
  gender: string;
  joinedAt: Date;
}