# lib/api-zod/src/generated/types/onlineUsersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { OnlineUser } from "./onlineUser";

export interface OnlineUsersResponse {
  users: OnlineUser[];
  total: number;
}