# lib/api-zod/src/generated/types/sendMessageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anonymous Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageRequest {
  sessionId: string;
  fromUserId: string;
  toUserId: string;
  fromUsername: string;
  /** @maxLength 1000 */
  content: string;
}